solucionado la separacion de responsabilidad en mappers de salida
