+++ logs/archived/users-service-2025-10-12.0.log
@@ -1,0 +1,11 @@
+2025-10-12 16:02:30 [restartedMain] INFO  c.d.l.LearncompanyApplication - Starting LearncompanyApplication using Java 17.0.16 with PID 987877 (/home/payan/Escritorio/Apps-empresariales/learn-company/build/classes/java/main started by payan in /home/payan/Escritorio/Apps-empresariales/learn-company)
+2025-10-12 16:02:30 [restartedMain] INFO  c.d.l.LearncompanyApplication - The following 1 profile is active: "dev"
+2025-10-12 16:02:34 [restartedMain] WARN  org.hibernate.orm.deprecation - HHH90000025: MariaDBDialect does not need to be specified explicitly using 'hibernate.dialect' (remove the property setting and it will be selected by default)
+2025-10-12 16:02:38 [restartedMain] WARN  o.s.b.a.o.j.JpaBaseConfiguration$JpaWebConfiguration - spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning
+2025-10-12 16:02:38 [restartedMain] WARN  o.s.b.a.s.s.UserDetailsServiceAutoConfiguration - 
+
+Using generated security password: a9c5b887-9e04-4b7f-8e66-7b8f86ac1296
+
+This generated password is for development use only. Your security configuration must be updated before running your application in production.
+
+2025-10-12 16:02:39 [restartedMain] INFO  c.d.l.LearncompanyApplication - Started LearncompanyApplication in 10.559 seconds (process running for 11.294)
